requirements.txt
absl-py==2.3.1
addict==2.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
basicsr==1.4.2
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
dotenv==0.9.9
facexlib==0.3.0
filelock==3.19.1
filterpy==1.4.5
Flask==3.1.1
fonttools==4.59.1
fsspec==2025.7.0
future==1.0.0
gfpgan==1.3.8
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.7.3
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.5
mod-wsgi==5.0.2
mpmath==1.3.0
networkx==3.5
ngrok==1.4.0
numba==0.61.2
numpy==2.2.6
openai==1.86.0
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.32.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.6
pydantic_core==2.33.2
pyhtml2pdf==0.0.7
pyngrok==7.2.11
pyodbc==5.2.0
pyparsing==3.2.3
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.25.2
scipy==1.16.1
selenium==4.34.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.14.0
tb-nightly==2.21.0a20250820
tensorboard-data-server==0.7.2
tifffile==2025.6.11
torch==2.0.0
torchvision==0.15.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yapf==0.43.0

# New dependencies for Document AI Platform v2.0
streamlit>=1.30.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-openai>=0.0.5
tiktoken>=0.5.2
python-Levenshtein>=0.21.0
fastapi>=0.95.0
uvicorn>=0.20.0
paddlepaddle>=2.5.0
paddleocr>=2.6.0
pytesseract>=0.3.10
pdf2image>=1.16.0
python-docx>=0.8.11
PyPDF2>=3.0.0
rake-nltk>=1.0.6
tenacity>=8.2.0

# Async task processing with Celery and Redis
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

# Testing and development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0
pre-commit>=3.6.0
bandit>=1.7.5
safety>=2.3.5
